Fix file names
