+++ AutomatedTesting/Gem/PythonTests/Physics/TestSuite_Periodic.py
@@ -401,20 +401,20 @@ class TestAutomation(TestAutomationBase):
         self._run_test(request, workspace, editor, test_module)
 
     @revert_physics_config
-    def test_Collider_SphereShapeEditting(self, request, workspace, editor, launcher_platform):
-        from .tests.collider import Collider_SphereShapeEditting as test_module
+    def test_Collider_SphereShapeEditing(self, request, workspace, editor, launcher_platform):
+        from .tests.collider import Collider_SphereShapeEditing as test_module
         self._run_test(request, workspace, editor, test_module,
                        extra_cmdline_args=["--regset=/Amazon/Preferences/EnablePrefabSystem=true"])
 
     @revert_physics_config
-    def test_Collider_BoxShapeEditting(self, request, workspace, editor, launcher_platform):
-        from .tests.collider import Collider_BoxShapeEditting as test_module
+    def test_Collider_BoxShapeEditing(self, request, workspace, editor, launcher_platform):
+        from .tests.collider import Collider_BoxShapeEditing as test_module
         self._run_test(request, workspace, editor, test_module,
                        extra_cmdline_args=["--regset=/Amazon/Preferences/EnablePrefabSystem=true"])
 
     @revert_physics_config
-    def test_Collider_CapsuleShapeEditting(self, request, workspace, editor, launcher_platform):
-        from .tests.collider import Collider_CapsuleShapeEditting as test_module
+    def test_Collider_CapsuleShapeEditing(self, request, workspace, editor, launcher_platform):
+        from .tests.collider import Collider_CapsuleShapeEditing as test_module
         self._run_test(request, workspace, editor, test_module,
                        extra_cmdline_args=["--regset=/Amazon/Preferences/EnablePrefabSystem=true"])
